Use custom models when defined
Everywhere in views, utils and probably other modules the concrete models are imported and used, which doesn't make sense if custom models have been registered.
